+++ app/src/pages/__generated__/SpanHeader_span.graphql.ts
@@ -1,5 +1,5 @@
 /**
- * @generated SignedSource<<6a4e67540b2600d845cf17eb8340841b>>
+ * @generated SignedSource<<fbe514afdb29c68bd3bf6b3f2fc069ff>>
  * @lightSyntaxTransform
  * @nogrep
  */
@@ -9,7 +9,7 @@
 // @ts-nocheck
 
 import { ReaderFragment } from 'relay-runtime';
-export type SpanKind = "agent" | "chain" | "embedding" | "evaluator" | "guardrail" | "llm" | "reranker" | "retriever" | "tool" | "unknown";
+export type SpanKind = "agent" | "chain" | "embedding" | "evaluator" | "guardrail" | "llm" | "prompt" | "reranker" | "retriever" | "tool" | "unknown";
 export type SpanStatusCode = "ERROR" | "OK" | "UNSET";
 import { FragmentRefs } from "relay-runtime";
 export type SpanHeader_span$data = {