+++ js/packages/phoenix-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @arizeai/phoenix-cli
 
+## 0.5.1
+
+### Patch Changes
+
+- @arizeai/phoenix-client@5.8.1
+
 ## 0.5.0
 
 ### Minor Changes