+++ js/packages/phoenix-otel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @arizeai/phoenix-otel
 
+## 0.4.0
+
+### Minor Changes
+
+- a8896db: feat: Enhance compatability with ai-sdk v6 telemetry
+
 ## 0.3.4
 
 ### Patch Changes